refactor: modularize knowledge API into organized package structure
Split the large knowledge.py file (1,029 lines) into a well-organized
package with focused modules for improved maintainability:

- api/knowledge/tools.py - Tool management and utility functions
- api/knowledge/core.py - Core knowledge CRUD operations
- api/knowledge/sync.py - Knowledge synchronization functionality
- api/knowledge/code_indexing.py - Code parsing (ctags/tree-sitter)
- api/knowledge/code_viewer.py - Code browsing and search
- api/knowledge/__init__.py - Package exports

Benefits:
✅ Improved modularity and separation of concerns
✅ Easier navigation and maintenance
✅ Better testability of individual components
✅ Full backward compatibility maintained
✅ All 31 knowledge API tests pass
